+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# WIP
+* Fix: Better Vivaldi detection for WebSocket (using speed dial and bookmark bar properties)
+
 # 2.0.2
 * Fix: Syntax error in Firefox